Update recommended extensions to include Puppet
This commit adds the `puppet.puppet-vscode` extension to the VS Code recommended extensions for this repo. If a user opens this repo with VS Code, it will prompt them to install the Puppet Extension, which will give them Bolt intellisense, hover information, and other features inside the Bolt plans. If a user does not have VS Code, this will not prompt them.