Add Bolt config file validation
This commit adds the `redhat.vscode-yaml` extension to the recommendations and configures it in the workspace settings file. This will prompt a user to install the Redhat YAML extension when they open this project in VS Code, and point the extension to the Bolt YAML schema files. This adds intellisense, hover, and validation support for the Bolt configuration files as well as Bolt YAML plans.